% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_pg.R
\name{connect_to_pg}
\alias{connect_to_pg}
\title{Connect to a PostgreSQL database}
\usage{
connect_to_pg()
}
\value{
DBI connection object
}
\description{
Prompts the user for a database profile, password, and schema.
Opens a connection using the connections pane and sets search_path.
}
