% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_pg.R
\name{disconnect_pg}
\alias{disconnect_pg}
\title{Disconnect from PostgreSQL and remove connection object}
\usage{
disconnect_pg(con)
}
\arguments{
\item{con}{A DBIConnection created by connect_pg()}
}
\description{
Disconnect from PostgreSQL and remove connection object
}
