% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{pg_profiles}
\alias{pg_profiles}
\title{Centralized PostgreSQL database profiles}
\usage{
pg_profiles()
}
\value{
List of profiles
}
\description{
Returns a named list of PostgreSQL connection parameters, including available schemas.
}
