% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_pg.R
\name{disconnect_pg}
\alias{disconnect_pg}
\title{Disconnect from PostgreSQL}
\usage{
disconnect_pg(con)
}
\description{
Works with both RStudio Connections pane connections
and plain DBI::dbConnect() connections.
}
