% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disconnect_pg.R
\name{disconnect_pg}
\alias{disconnect_pg}
\title{Disconnect from PostgreSQL}
\usage{
disconnect_pg(con)
}
\description{
Works with connections created using the RStudio connections pane
(class "connConnection") and also raw DBI connections.
}
