% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pq_enum.R
\name{is_pg_enum}
\alias{is_pg_enum}
\title{Check if string is a PostgreSQL enum}
\usage{
is_pg_enum(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
Logical vector
}
\description{
Returns TRUE if a string starts with 'pq_', otherwise FALSE.
}
